requirements.txt
httpx
pydantic
python-dotenv
fastapi
uvicorn
tiktoken